Add headline and tweet processing operations
In this commit, add a new table to house the processed data. The
processed data is simply headlines and tweets resampled into 30-minute
intervals, with their sentiment values being averaged. These values for
headlines and tweets are then merged on the resampled timestamps.

Supporting endpoint and crud operations are included.